MOD: Pass a 'reference window' pointer to RenderedString related functions.
This change allows RenderedString and related components access to the window
the RenderedString is intended to be rendered for (this can be 0, and
components are required to do something sane (or at least not crash).

This also fixes the [window] tag supported by the default parser.
See: http://www.cegui.org.uk/mantis/view.php?id=509