Added support for a read-only mouse cursor image to Editboxes.
Patch v2 thanks to the hints and advice from Ident. :)

Feature:
Editboxes now have a 'ReadOnlyMouseCursorImage' property
that will be used to set the cursor image when the edit box is read only.
Note: The default mouse cursor is still used when the property is empty,
and maybe that's not what we want.


Changes from patch v1:
- Against branch v0 instead of default.
- Doesn't overwrite the MouseCursorImage property anymore,
  as this property may be needed after being set read-only.
- I didn't use CEGUI_DEFINE_WINDOW_RENDERER_PROPERTY even
  if advised to do so, because this property can only work
  with the FalagardEditoBox class for instance, and I didn't know
  how to be able to call something like FalagardEditobx::editboxSetReadOnly()
  from the Editbox class.

The change has been tested in OpenDungeons and is working fine. :)

Feel free to comment.