Use a workaround around a bug in some builds of MinGW-w64 that sometimes gives a false warning.
When using a function from the "printf"/"scanf" family with a format containing "%llu", we might get a warning saying that this format isn't recognized. However it does work fine, and the bug is in the warning being issued. This bug happens only sometimes and in some builds of MinGW-w64. I've used a nasty work-around to avoid those false warnings.