Add to class "Element" 2 new properties "AdjustWidthToContent" and "AdjustHeightToContent".
If "AdjustWidthToContent" is "true", keep the width of the element at the minimal value in which the whole element content is visible without the need for a horizontal scrollbar (if possible), and while the content remains "intact" (if possible).

The meaning of "element content" and "intact" depends on the type of element. For example, for a "FalagardStaticText" widget, the content is the text, and "intact" means no single word is split between 2 or more lines. So setting this property to "true" would keep the width of the widget to the minimal value in which the whole text is visible without the need for a horizontal scrollbar (if possible), and without the need to split a word between 2 or more lines (if possible).

The behaviour of "AdjustHeightToContent" is similar, affecting the element's height rather than its width.

Whenever the size of the widget should be (re)adjusted to its content, this is done by calling "adjustSizeToContent". This happens in the default implementations of "onIsSizeAdjustedToContentChanged" and "onSized". The cegui user must nake sure to call "adjustSizeToContent" whenever the size should be re-adjusted to the content in any other case.

The behaviour of "adjustSizeToContent" is as follows: if the "AdjustWidthToContent" property is set to "true", the width is adjusted. If the "AdjustHeightToContent" property is set to "true", the height is adjusted. The exact behaviour of this method depends on the combination of "AdjustWidthToContent" and "AdjustHeightToContent", as well as on the type of widget, and possibly other properties. You might want to look at the documentation for "FalagardStaticText::adjustSizeToContent" (which will be added in a later commit in this PR) for a detailed description on how this method behaves for a "FalagardStaticText" widget, which serves as a great example.

There are 2 helper methods (which will be added in a later commit in this PR) that the cegui user may want to use when implementing this method:
1. "adjustSizeToContent_direct"
2. "getSizeAdjustedToContent_bisection"

The default implementation of "adjustSizeToContent" calls "adjustSizeToContent_direct".

Also add an event "Element::IsSizeAdjustedToContentChanged" fired when any of the properties "AdjustWidthToContent" and "AdjustHeightToContent" changes.