add popup menu positioning logic (to avoid menus being clipped)
When using the built-in MenuBar, MenuItem, and PopupMenu widgets
together, CEGUI provides considerable built-in functionality to
behave like "typical" OS menus, i.e. opening the menus when clicked
on, closing when clicking away, etc.

However it doesn't handle the scenario when the popup menu would
be off the screen. On most OS's there is some logic so that the
menu will be placed differently to avoid that when it would happen,
since it can be frustrating for the user.

In this commit we add such logic and make it the default behavior.
It has the following steps, when a menu item is a child of a
menubar or popupmenu:

- Find the rectangle within which the popup menu must appear to avoid
  clipping.
- There are two candidate offsets for menubars and popupmenus:
  In menubars, we use a vertical configuration, either above or
  below the menu item. In popupmenus, we place the submenu either to
  the left or the right of the menu item.
- For each candidate offset, compute how it must be adjusted to avoid
  clipping.
- Choose the primary offset, unless the secondary candidate is a better
  fit.

Both the clipping rectangle determination, and the final popup
position determination, are implemented in virtual functions, so
that users of the library can override them to implement this
functionality differently.

The manual "popup offset" member is removed as well, as per our
discussions in PR #231

https://bitbucket.org/cegui/cegui/pull-requests/231