add a MenubarDirection enum, property to MenuBase and Menubar
This is used to determine where popup menus from a menubar are placed,
the options are down, up, and a "best fit" strategy which is similar
to that used for submenus.

In this commit, Down is the default value of this option since it's
the most backwards compatible, and there were some questions raised
in code review as to how to actual implement a good "best fit"
strategy.

In comments on the PR it was suggested that actually, rather than one
clipping rectangle, we should use two rectangles -- both the overall
clipping rectangle for the root container, and the parent of the menubar.

Roughly, the outer rectangle is like a "hard" limit which must be
respected, and the inner rectangle is like a "soft" limit or a "hint"
which is used to help decide whether to open up or down.

I think this is interesting, and there were some compelling scenarios
described in the PR comments about why this might be a good idea. But
it's also more complex, and I'm basically not 100% comfortable with it
so I'm inclined to think about it some more before implementing it. It's
not totally clear to me what is the best way to take into account the
"hint" rectangle versus the "clip" rectangle, so as a result I feel like
I can't make a good argument why there should be two rectangles. At the
same time, I agree that my original idea has flaws.

So I don't implement that in this commit. But we do add a tristate
option for menubars, which we agreed was good regardless.

It might be that we should simply have several "fit" strategies, I'm
not sure though. Many users are surely going to overdose on this
level of detail anyways :D