Fix debug build and rework Exceptions includes logic
Commit e18e48624cd4 broke the debug build. HAVE_EXECINFO_H is never
defined by the build system so this header is never included, even
when it is needed by the debug code.

The right way to fix this is only to include these headers when debug
is actually enabled. This does mean that the debug build is still
broken under musl but fixing that would require using libunwind
instead.